add isSelected param to facultyUser and logic to endpoints
