+++ IzvorniKod/src/main/java/com/educhat/backend/models/FacultyUser.java
@@ -25,4 +25,7 @@ public class FacultyUser {
 
     @NotNull
     private boolean kicked;
+
+    @NotNull
+    private boolean isSelected;
 }
