+++ IzvorniKod/src/main/java/com/educhat/backend/repository/FacultyUserRepository.java
@@ -2,8 +2,6 @@ package com.educhat.backend.repository;
 
 import com.educhat.backend.models.FacultyUser;
 import com.educhat.backend.models.enums.Role;
-import jakarta.validation.constraints.NotNull;
-import lombok.NonNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -13,8 +11,6 @@ import java.util.Optional;
 
 @Repository
 public interface FacultyUserRepository extends JpaRepository<FacultyUser, Long> {
-
-    boolean existsByUserIdAndFacultyIdAndRole(Long userId, Long facultyId, Role role);
 
     List<FacultyUser> findByUserIdAndRole(Long userId, Role role);
