+++ IzvorniKod/src/main/java/com/educhat/backend/services/UserService.java
@@ -103,81 +103,4 @@ public class UserService {
         return new UserDetailsDTO(user.getRealUsername(), user.getEmail(), numFaculties, numPosts, numLikes);
     }
 
-    public List<FacultyUser> createFacultyUser(Long userId, List<FacultyUserCreateDTO> facultyUserCreateDTOs) {
-        // check if user exists
-        if(!userRepository.existsById(userId)) {
-            throw new UserNotFoundException("User not found");
-        }
-
-        // list of all saved faculty users
-        List<FacultyUser> savedFacultyUsers = new ArrayList<>();
-
-        // iterate through all selected faculties
-        for(FacultyUserCreateDTO dto : facultyUserCreateDTOs) {
-
-            // check if faculty exists
-            if(!facultyRepository.existsById(dto.getFacultyId())) {
-                throw new FacultyNotFoundException("Faculty not found");
-            }
-
-            // already exist that faculty user
-            Optional<FacultyUser> facultyUserOptional = facultyUserRepository.findByUserIdAndFacultyId(userId, dto.getFacultyId());
-            if(facultyUserOptional.isPresent()) {
-                FacultyUser facultyUser = facultyUserOptional.get();
-                // if the role is different, update it
-                if(!facultyUser.getRole().equals(dto.getUserRole())) {
-                    facultyUser.setRole(dto.getUserRole());
-                    facultyUserRepository.save(facultyUser);
-                    savedFacultyUsers.add(facultyUser);
-                }
-            // create new faculty user
-            } else {
-                FacultyUser facultyUser = new FacultyUser();
-                facultyUser.setUserId(userId);
-                facultyUser.setFacultyId(dto.getFacultyId());
-                facultyUser.setRole(dto.getUserRole());
-                facultyUser.setKicked(false);
-
-                facultyUserRepository.save(facultyUser);
-                savedFacultyUsers.add(facultyUser);
-            }
-        }
-
-        return savedFacultyUsers;
-    }
-
-    public List<SelectedFacultiesDTO> selectedFaculties(Long userId) {
-        // find user if exist
-        User user = userRepository.findById(userId)
-                .orElseThrow( () -> new UserNotFoundException("User not found"));
-
-        // initialize empty list - zero selected faculties
-        List<SelectedFacultiesDTO> selectedFaculties = new ArrayList<>();
-
-        // iterate through faculties user selected
-        for(FacultyUser facultyUser : facultyUserRepository.findByUserId(userId)) {
-            // if user is kicked from that faculty, don't show it
-            if(facultyUser.isKicked()) continue;
-
-            // find faculty
-            Faculty faculty = facultyRepository.findById(facultyUser.getFacultyId())
-                    .orElseThrow( () -> new FacultyNotFoundException("Faculty not found"));
-
-            // find associated faculty years
-            List<FacultyYearDTO> facultyYearDTOs = new ArrayList<>();
-            for(FacultyYear facultyYear : facultyYearRepository.findByFacultyId(faculty.getId())) {
-                facultyYearDTOs.add(new FacultyYearDTO(facultyYear.getId(), facultyYear.getTitle()));
-            }
-
-            // add created dto to list
-            selectedFaculties.add(new SelectedFacultiesDTO(
-                    faculty.getId(),
-                    faculty.getTitle(),
-                    facultyYearDTOs,
-                    user.getRole().equals(Role.ADMIN),
-                    facultyUser.getRole().equals(Role.ADMIN) || user.getRole().equals(Role.ADMIN)));
-        }
-
-        return selectedFaculties;
-    }
 }
